mesh, shaders: Use uvec4 for joints vertex attribute
Currently, the joints attribute is passed in as a vector of 4 bytes that
glVertexAttribPointer() converts into vec4, which the shaders expect and
typecase to int when it's used. This is also how GLTF supplies joints.

At the same time, Metal doesn't do implicit conversions like that; its
shaders are transpiled from GLSL, which doesn't have a type for a vector
of bytes, so the resulting MSL shaders expect Float4, resulting in
disfigured skinning.

Solve this by explicitly converting joints to an integer array on load
and change corresponding mesh attribute type to DT_UVEC4, while at the
same time changing the shaders' joints attributes to uvec4. This works
correctly with both render backends.

+++ shaders/model.vert
@@ -6,7 +6,7 @@ layout (location=ATTR_LOC_POSITION) in vec3 position;
 layout (location=ATTR_LOC_TEX) in vec2 tex;
 layout (location=ATTR_LOC_NORMAL) in vec3 normal;
 layout (location=ATTR_LOC_TANGENT) in vec4 tangent;
-layout (location=ATTR_LOC_JOINTS) in vec4 joints;
+layout (location=ATTR_LOC_JOINTS) in uvec4 joints;
 layout (location=ATTR_LOC_WEIGHTS) in vec4 weights;
 
 #include "ubo_lighting.glsl"
@@ -33,7 +33,7 @@ void main()
     vec4 total_normal = vec4(0, 0, 0, 0);
     if (use_skinning != 0) {
         for (int i = 0; i < 4; i++) {
-            mat4 joint_transform = joint_transforms[int(joints[i])];
+            mat4 joint_transform = joint_transforms[joints[i]];
             vec4 local_pos = joint_transform * vec4(position, 1.0);
             total_local_pos += local_pos * weights[i];
